+++ overlord_test/test.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"encoding/json"
+	"flag"
 	"fmt"
 	"log"
 	"os"
@@ -100,6 +101,8 @@ func runTests(config *util.HeatConfig, details *util.StackDetails) {
 }
 
 func main() {
+	flag.Parse()
+
 	params := map[string]string{
 		"git_command": createGitCmdParam(),
 	}
@@ -109,5 +112,8 @@ func main() {
 
 	config, stackDetails := framework.BuildConfigAndCreateStack(&params)
 	runTests(config, stackDetails)
-	goheat.DeleteStack(config, stackDetails.Stack.Links[0].Href)
+
+	if *framework.DeleteStack {
+		goheat.DeleteStack(config, stackDetails.Stack.Links[0].Href)
+	}
 }
